refactor : 마이페이지 중 내 정보 변경에 해당하는 API 병합 (#196)
